backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
pydantic
email-validator
pydantic-settings